services/sockets/socket.requirements.txt
fastapi
uvicorn
websockets
python-jose
pydantic
python-multipart
PyJWT==1.7.1
docker
aiohttp
pika